+++ src/example_app/example_app/documents/admin.py
@@ -5,7 +5,7 @@
 """
 from restfw_admin import widgets as all_widgets
 from restfw_admin.config import resource_admin_config
-from restfw_admin.resource_admin import Exclude, ResourceAdmin, ViewSettings
+from restfw_admin.resource_admin import Exclude, Only, ResourceAdmin, ViewSettings
 from restfw_admin.validators import Required
 from .views import DocView, DocsView
 
@@ -17,15 +17,28 @@ class DocsAdmin(ResourceAdmin):
     child_view_class = DocView
     location = '/docs'
     index = 1
+    default_fields = Exclude(
+        '_embedded'
+    )
+    fields = Only(
+        '_links.self.href',
+        'id',
+        'user_id',
+        'data',
+        'meta',
+    )
     list_view = ViewSettings(
-        fields=Exclude('data'),
+        fields=Exclude(
+            'data',
+            'meta',
+        ),
         widgets={
             'user_id': all_widgets.ReferenceField(
                 reference='users',
                 reference_field='first_name',
                 label='User',
                 link='show',
-            )
+            ),
         }
     )
     show_view = ViewSettings(
@@ -37,6 +50,7 @@ class DocsAdmin(ResourceAdmin):
                 link='show',
             ),
             'data': all_widgets.RichTextField(),
+            'meta': all_widgets.JsonField(),
         }
     )
     create_view = ViewSettings(
@@ -48,10 +62,12 @@ class DocsAdmin(ResourceAdmin):
                 validators=[Required()]
             ),
             'data': all_widgets.RichTextInput(),
+            'meta': all_widgets.JsonInput(),
         }
     )
     edit_view = ViewSettings(
         widgets={
             'data': all_widgets.RichTextInput(),
+            'meta': all_widgets.JsonInput(),
         }
     )