+++ src/example_app/example_app/documents/schemas.py
@@ -13,6 +13,7 @@ class DocSchema(schemas.HalResourceSchema):
     id = schemas.UnsignedIntegerNode(title='ID')
     user_id = schemas.IntegerNode(title='User ID')
     data = schemas.EmptyStringNode(title='Document data')
+    meta = schemas.MappingNode(title='Meta', unknown='preserve')
 
 
 class DocsSchema(schemas.HalResourceWithEmbeddedSchema):
@@ -28,6 +29,7 @@ class DocsSchema(schemas.HalResourceWithEmbeddedSchema):
 class CreateDocSchema(schemas.MappingSchema):
     user_id = schemas.IntegerNode(title='User ID', validator=user_id_validator)
     data = schemas.EmptyStringNode(title='Document data')
+    meta = schemas.MappingNode(title='Meta', unknown='preserve')
 
 
 PatchDocSchema = schemas.clone_schema_class(